Some of the interface bugs fixed, edit transaction (almost) finalized
I created a new bug with a dropdown list at editing transaction, but in general it works. Also fixed 'Amount' column in database, added some forms validations for trip and transaction names

+++ database.py
@@ -35,7 +35,7 @@ class Transaction(Base):
     __tablename__ = 'transactions'
     trans_id = Column(Integer, primary_key=True, index = True, autoincrement=True)
     tr_name = Column(String)
-    amount = Column(BigInteger)
+    amount = Column(Float)
     trip_id = Column(BigInteger, ForeignKey('trips.trip_id'))
 
     trip = relationship("Trip", back_populates="transactions")
